chore: move `AlreadyBuiltWheelError` & `NonPlatformWheelError` to `errors.py`
