+++ cibuildwheel/linux.py
@@ -18,10 +18,8 @@ from .oci_container import OCIContainer, OCIContainerEngineConfig
 from .options import BuildOptions, Options
 from .typing import PathOrStr
 from .util import (
-    AlreadyBuiltWheelError,
     BuildFrontendConfig,
     BuildSelector,
-    NonPlatformWheelError,
     find_compatible_wheel,
     get_build_verbosity_extra_flags,
     prepare_command,
@@ -306,7 +304,7 @@ def build_in_container(
             container.call(["mkdir", "-p", repaired_wheel_dir])
 
             if built_wheel.name.endswith("none-any.whl"):
-                raise NonPlatformWheelError()
+                raise errors.NonPlatformWheelError()
 
             if build_options.repair_command:
                 log.step("Repairing wheel...")
@@ -321,7 +319,7 @@ def build_in_container(
 
             for repaired_wheel in repaired_wheels:
                 if repaired_wheel.name in {wheel.name for wheel in built_wheels}:
-                    raise AlreadyBuiltWheelError(repaired_wheel.name)
+                    raise errors.AlreadyBuiltWheelError(repaired_wheel.name)
 
         if build_options.test_command and build_options.test_selector(config.identifier):
             log.step("Testing wheel...")
