+++ cibuildwheel/pyodide.py
@@ -17,10 +17,8 @@ from .options import Options
 from .typing import PathOrStr
 from .util import (
     CIBW_CACHE_PATH,
-    AlreadyBuiltWheelError,
     BuildFrontendConfig,
     BuildSelector,
-    NonPlatformWheelError,
     call,
     combine_constraints,
     download,
@@ -299,7 +297,7 @@ def build(options: Options, tmp_path: Path) -> None:
                 built_wheel = next(built_wheel_dir.glob("*.whl"))
 
                 if built_wheel.name.endswith("none-any.whl"):
-                    raise NonPlatformWheelError()
+                    raise errors.NonPlatformWheelError()
 
                 if build_options.repair_command:
                     log.step("Repairing wheel...")
@@ -316,7 +314,7 @@ def build(options: Options, tmp_path: Path) -> None:
                 repaired_wheel = next(repaired_wheel_dir.glob("*.whl"))
 
                 if repaired_wheel.name in {wheel.name for wheel in built_wheels}:
-                    raise AlreadyBuiltWheelError(repaired_wheel.name)
+                    raise errors.AlreadyBuiltWheelError(repaired_wheel.name)
 
             if build_options.test_command and build_options.test_selector(config.identifier):
                 log.step("Testing wheel...")
